+++ packages/block-editor/src/hooks/index.js
@@ -1,12 +1,16 @@
 /**
  * Internal dependencies
  */
-import { createBlockEditFilter, createBlockListBlockFilter } from './utils';
+import {
+	createBlockEditFilter,
+	createBlockListBlockFilter,
+	createBlockSaveFilter,
+} from './utils';
 import './compat';
 import align from './align';
 import './lock';
 import anchor from './anchor';
-import './aria-label';
+import ariaLabel from './aria-label';
 import customClassName from './custom-class-name';
 import './generated-class-name';
 import style from './style';
@@ -19,7 +23,7 @@ import border from './border';
 import position from './position';
 import layout from './layout';
 import childLayout from './layout-child';
-import './content-lock-ui';
+import contentLockUI from './content-lock-ui';
 import './metadata';
 import customFields from './custom-fields';
 import blockHooks from './block-hooks';
@@ -34,6 +38,7 @@ createBlockEditFilter(
 		duotone,
 		position,
 		layout,
+		contentLockUI,
 		window.__experimentalConnections ? customFields : null,
 		blockHooks,
 		blockRenaming,
@@ -49,6 +54,17 @@ createBlockListBlockFilter( [
 	border,
 	position,
 	childLayout,
+] );
+createBlockSaveFilter( [
+	align,
+	anchor,
+	ariaLabel,
+	customClassName,
+	border,
+	color,
+	style,
+	fontFamily,
+	fontSize,
 ] );
 
 export { useCustomSides } from './dimensions';