+++ packages/commands/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Enhancements
+
+-   Support conditional commands and commands loaders using the "disabled" config.
+
 ## 0.19.0 (2023-12-13)
 
 ## 0.18.0 (2023-11-29)